requirements.txt
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.22.0
django-cors-headers==3.13.0
sendgrid==6.9.7
psycopg2-binary==2.9.4
mailjet-rest==1.3.4